chore: remove command.test.ts file, split this file into separate file by command and add to vitest config pool forks
