+++ vitest.config.ts
@@ -31,6 +31,7 @@ const configExtension: { [key: string]: ViteUserConfig } = {
   e2e: defineConfig({
     test: {
       include: ['tests/e2e/**/*.test.ts'],
+      pool: 'forks',
     },
   }),
   'smoke-rebilly': defineConfig({
